requirements.txt
ephem==4.1.5
etcd3==0.12.0
grpcio==1.66.1
protobuf==3.20.1
PyYAML==6.0.2
six==1.16.0
tenacity==9.0.0
